+++ frontend/src/views/main-capacity/capacity-detail.js
